refactoring in progress, unit tests refactoring in progress
