+++ lib/helpers/error.js
