+++ lib/helpers/stringify.js
