+++ lib/highlogger.js
@@ -3,9 +3,9 @@
 let instance,
     Transporters = require('./transporters'),
     Debug = require('./debug'),
-    constants = require('../helpers/constants'),
+    constants = require('./helpers/constants'),
     path = require('path'),
-    error = require('../helpers/error');
+    error = require('./helpers/error');
 
 const EMPTY = '';
