+++ lib/transporters.js
@@ -3,9 +3,9 @@
 let path = require('path'),
     glob = require('glob'),
     async = require('async'),
-    transporterFiles = glob.sync('../transporters/*.js', {cwd: __dirname}),
-    constants = require('../helpers/constants'),
-    error = require('../helpers/error');
+    transporterFiles = glob.sync('./transporters/*.js', {cwd: __dirname}),
+    constants = require('./helpers/constants'),
+    error = require('./helpers/error');
 
 class Transporters {
