+++ lib/transporters/abstract.js
