+++ lib/transporters/console.js
