+++ lib/transporters/socket.js
