+++ lib/transporters/syslog.js
