+++ test/transporter/syslog-test.js
@@ -1,7 +1,7 @@
 'use strict';
 
-let SocketTransporter = require('../../transporters/socket'),
-    SyslogTransporter = require('../../transporters/syslog'),
+let SocketTransporter = require('../../lib/transporters/socket'),
+    SyslogTransporter = require('../../lib/transporters/syslog'),
     moment = require('moment'),
     parseFormat = require('moment-parseformat'),
     assert = require('assert'),
@@ -9,7 +9,7 @@ let SocketTransporter = require('../../transporters/socket'),
     fs = require('fs'),
     dgram = require('dgram');
 
-const SHARED_CONSTANTS = require('../../helpers/constants');
+const SHARED_CONSTANTS = require('../../lib/helpers/constants');
 
 describe('transporter syslog', function () {
   describe('init', function () {
